http://puu.sh/1eWN9 > Chatrooms and Messages
Made controllers for chatrooms and messages so you can make new
chatrooms and messages and have them displayed. No user controllers or
models are made yet, so the create controller for message just makes the
author Bob.

Haven't put in any html.erb's yet.

There's a table I can't get rid of in the schema, called comments which
I accidentaly put in. Don't know if it'll be there for you too. (Is the
database stored in the UWAnnachat folder?)